9. Площадь треугольника. [#464592]

+++ src/ru/job4j/condition/SqArea.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(double p, double k) {
+    public static double square(int p, int k) {
         double h = p / (2 * (k + 1)) ;
         double L = h * k;
         double S = L * h;
